+++ lib/very_commons.dart
@@ -4,5 +4,7 @@ export 'src/limit.dart';
 export 'src/page.dart';
 export 'src/page_request.dart';
 export 'src/pageable.dart';
+export 'src/scroll_position.dart';
 export 'src/slice.dart';
 export 'src/sort.dart';
+export 'src/window.dart';